+++ dash/utils/cache_utils.py
@@ -26,7 +26,7 @@ logger=logging.getLogger('adesit_callbacks')
 cache = None
 
 default_data = {
-    "data_holder": None,
+    'data_holder': None,
     "graphs": {},
     "thresolds_settings": {},
     "table_data": None,
@@ -41,12 +41,6 @@ def find_resolution_of_attribute(col):
     diff = scol[1:]-scol[:-1]
     return diff.min()
 
-def strip_to_id(attr_name):
-    for char in ['.', '{', '}']:
-        attr_name = attr_name.replace(char, '_')
-    attr_name = attr_name.lower()
-    return attr_name
-
 def gen_data_holder(df):
     # Proprocessing dataframe
     df = df.dropna()
@@ -54,8 +48,6 @@ def gen_data_holder(df):
     cols_minmax = {}
     cols_ncats = {}
     cols_res = {}
-    time_cols = []
-    cols_ids = {}
     for c in df.columns:
         if c in ['id', 'Id', 'ID']:
             df = df.drop(columns=c)
@@ -63,10 +55,8 @@ def gen_data_holder(df):
             try:
                 cols_type[c] = DATETIME_COLUMN
                 df[c] = pd.to_datetime(df[c], infer_datetime_format=True)
-                time_cols.append(c)
                 cols_minmax[c] = [df[c].min(), df[c].max()]
                 cols_res[c] = find_resolution_of_attribute(df[c])
-                cols_ids[c] = strip_to_id(c)
             except ValueError:
                 cols_type[c] = CATEGORICAL_COLUMN
                 cols_res[c] = 1
@@ -77,12 +67,10 @@ def gen_data_holder(df):
                     "label_encoder": le,
                 }
                 cols_minmax[c] = [0, len(le.classes_)-1]
-                cols_ids[c] = strip_to_id(c)
         elif is_numeric_dtype(df[c]):
             cols_type[c] = NUMERICAL_COLUMN
             cols_minmax[c] = [df[c].min(), df[c].max()]
             cols_res[c] = find_resolution_of_attribute(df[c])
-            cols_ids[c] = strip_to_id(c)
         else:
             df = df.drop(columns=c)
 
@@ -94,13 +82,11 @@ def gen_data_holder(df):
         "data": df,
         "full_data": df,
         "graph": None,
-        "user_columns": cols,
-        "columns_type": cols_type,
+        'user_columns': cols,
+        'columns_type': cols_type,
         "columns_minmax": cols_minmax,
         "columns_resolution": cols_res,
-        "columns_stripped_id": cols_ids,
         "categorical_columns_infos": cols_ncats,
-        "time_columns": time_cols,
         "time_infos": None,
         "X": [],
         "Y": []
@@ -136,7 +122,7 @@ def get_data(session_id, pydata=False, clear=False, filename=None, contents=None
             return None
 
         final_data = default_data
-        final_data["data_holder"]=gen_data_holder(df)
+        final_data['data_holder']=gen_data_holder(df)
         return final_data
 
     if clear:
